requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
pymysql==1.1.0
pandas==2.2.0
openpyxl==3.1.2
python-dotenv==1.0.0
pydantic==2.5.0
croniter==2.0.1
requests==2.31.0
kafka-python==2.2.15
